ChatServerMT 클린코드

+++ 01 Network/ChatServerMT/Sector.h
@@ -6,8 +6,8 @@
 
 static struct Sector {
 public:
-	int x;
-	int y;
+	short x;
+	short y;
 
 public:
 	bool operator==(const Sector& other);
